pyproject.toml
[tool.poetry]
name = "data_engineering"
version = "0.1.0"
description = ""
authors = ["Relativity74205 <arkadius@schuchhardt.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.3.0"
notebook = "^6.4.0"
psycopg2-binary = "^2.9.1"
jupyter = "^1.0.0"
cassandra-driver = "^3.25.0"
boto3 = "^1.18.21"
pyspark = "^3.1.2"
jupyterlab = "^3.1.14"
pyarrow = "^5.0.0"
toml = "^0.10.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
